refactor: rename ignore source files to copy-filter
