+++ src/copy-filter.js
